+++ world/src/main/kotlin/wumo/sim/envs/toy_text/DiscreteEnv.kt
@@ -5,18 +5,18 @@ import wumo.sim.spaces.Discrete
 import wumo.sim.util.argmax
 import wumo.sim.util.tuple4
 
-typealias Transition = Array<Array<MutableList<tuple4<Double, Int, Double, Boolean>>>>
+typealias Transition = Array<Array<MutableList<tuple4<Float, Int, Float, Boolean>>>>
 
 abstract class DiscreteEnv(val nS: Int,
                            val nA: Int,
                            val P: Transition,
-                           val isd: DoubleArray) : Env<Int, Int> {
+                           val isd: FloatArray) : Env<Int, Int> {
   var lastAction = -1
   var s = -1
   override val action_space = Discrete(nA)
   override val observation_space = Discrete(nS)
   
-  override fun step(a: Int): tuple4<Int, Double, Boolean, Map<String, Any>> {
+  override fun step(a: Int): tuple4<Int, Float, Boolean, Map<String, Any>> {
     val transitions = P[s][a]
     val i = argmax(0..transitions.lastIndex) { transitions[it]._1 }
     val (p, s, r, d) = transitions[i]
