+++ world/src/test/kotlin/wumo/sim/world/examples/algorithm/LinearTileCodingFunc.kt
@@ -3,8 +3,8 @@ package wumo.sim.world.examples.algorithm
 import wumo.sim.util.ndarray.NDArray
 
 class LinearTileCodingFunc(val feature: SuttonTileCoding) {
-  val w = DoubleArray(feature.numOfComponents)
-  operator fun invoke(s: NDArray<Double>, a: Int): Double {
+  val w = FloatArray(feature.numOfComponents)
+  operator fun invoke(s: NDArray<Float>, a: Int): Float {
     return w.innerProduct(feature(s, a))
   }
 }
